streamlit_test_app/requirements.txt
numpy==1.21.2
pandas==1.3.3
streamlit==0.85.0
Pillow==8.3.2
scikit-learn==1.0.1
